+++ packages/web3-core/src/AbstractWeb3Module.js
@@ -50,7 +50,7 @@ export default class AbstractWeb3Module {
         this._defaultAccount = options.defaultAccount ? toChecksumAddress(options.defaultAccount) : undefined;
         this._defaultBlock = options.defaultBlock || 'latest';
         this._transactionBlockTimeout = options.transactionBlockTimeout || 50;
-        this._transactionConfirmationBlocks = options.transactionConfirmationBlocks || 24;
+        this._transactionConfirmationBlocks = options.transactionConfirmationBlocks || 0;
         this._transactionPollingTimeout = options.transactionPollingTimeout || 750;
         this._defaultGasPrice = options.defaultGasPrice;
         this._defaultGas = options.defaultGas;