+++ packages/web3-eth-ens/src/contracts/Registry.js
@@ -44,6 +44,231 @@ export default class Registry extends AbstractContract {
     }
 
     /**
+     * TODO: Remove setter
+     *
+     * Setter for the transactionSigner property
+     *
+     * @property transactionSigner
+     *
+     * @param {TransactionSigner} value
+     */
+    set transactionSigner(value) {
+        if (this.resolverContract) {
+            this.resolverContract.transactionSigner = value;
+        }
+
+        super.transactionSigner = value;
+    }
+
+    /**
+     * Getter for the transactionSigner property
+     *
+     * @property transactionSigner
+     *
+     * @returns {TransactionSigner}
+     */
+    get transactionSigner() {
+        return this._transactionSigner;
+    }
+
+    /**
+     * Clears all subscriptions and listeners
+     *
+     * @method clearSubscriptions
+     *
+     * @returns {Promise<Boolean|Error>}
+     */
+    clearSubscriptions() {
+        if (this.resolverContract) {
+            this.resolverContract.clearSubscriptions('eth_unsubscribe');
+        }
+
+        return super.clearSubscriptions('eth_unsubscribe');
+    }
+
+    /**
+     * Sets the defaultGasPrice property on all contracts and on all sub-modules
+     *
+     * @property defaultGasPrice
+     *
+     * @param {String|Number} value
+     */
+    set defaultGasPrice(value) {
+        if (this.resolverContract) {
+            this.resolverContract.defaultGasPrice = value;
+        }
+
+        super.defaultGasPrice = value;
+    }
+
+    /**
+     * Gets the defaultGasPrice property
+     *
+     * @property defaultGasPrice
+     *
+     * @returns {String|Number} value
+     */
+    get defaultGasPrice() {
+        return super.defaultGasPrice;
+    }
+
+    /**
+     * Sets the defaultGas property on all contracts and on all sub-modules
+     *
+     * @property defaultGas
+     *
+     * @param {Number} value
+     */
+    set defaultGas(value) {
+        if (this.resolverContract) {
+            this.resolverContract.defaultGas = value;
+        }
+
+        super.defaultGas = value;
+    }
+
+    /**
+     * Gets the defaultGas property
+     *
+     * @property defaultGas
+     *
+     * @returns {String|Number} value
+     */
+    get defaultGas() {
+        return super.defaultGas;
+    }
+
+    /**
+     * Sets the transactionBlockTimeout property on all contracts and on all sub-modules
+     *
+     * @property transactionBlockTimeout
+     *
+     * @param {Number} value
+     */
+    set transactionBlockTimeout(value) {
+        if (this.resolverContract) {
+            this.resolverContract.transactionBlockTimeout = value;
+        }
+
+        super.transactionBlockTimeout = value;
+    }
+
+    /**
+     * Gets the transactionBlockTimeout property
+     *
+     * @property transactionBlockTimeout
+     *
+     * @returns {Number} value
+     */
+    get transactionBlockTimeout() {
+        return super.transactionBlockTimeout;
+    }
+
+    /**
+     * Sets the transactionConfirmationBlocks property on all contracts and on all sub-modules
+     *
+     * @property transactionConfirmationBlocks
+     *
+     * @param {Number} value
+     */
+    set transactionConfirmationBlocks(value) {
+        if (this.resolverContract) {
+            this.resolverContract.transactionConfirmationBlocks = value;
+        }
+
+        super.transactionConfirmationBlocks = value;
+    }
+
+    /**
+     * Gets the transactionConfirmationBlocks property
+     *
+     * @property transactionConfirmationBlocks
+     *
+     * @returns {Number} value
+     */
+    get transactionConfirmationBlocks() {
+        return super.transactionConfirmationBlocks;
+    }
+
+    /**
+     * Sets the transactionPollingTimeout property on all contracts and on all sub-modules
+     *
+     * @property transactionPollingTimeout
+     *
+     * @param {Number} value
+     */
+    set transactionPollingTimeout(value) {
+        if (this.resolverContract) {
+            this.resolverContract.transactionPollingTimeout = value;
+        }
+
+        super.transactionPollingTimeout = value;
+    }
+
+    /**
+     * Gets the transactionPollingTimeout property
+     *
+     * @property transactionPollingTimeout
+     *
+     * @returns {Number} value
+     */
+    get transactionPollingTimeout() {
+        return super.transactionPollingTimeout;
+    }
+
+    /**
+     * Sets the defaultAccount property on all contracts and on the personal module
+     *
+     * @property defaultAccount
+     *
+     * @param {String} value
+     */
+    set defaultAccount(value) {
+        if (this.resolverContract) {
+            this.resolverContract.defaultAccount = value;
+        }
+
+        super.defaultAccount = value;
+    }
+
+    /**
+     * Gets the defaultAccount property
+     *
+     * @property defaultAccount
+     *
+     * @returns {String} value
+     */
+    get defaultAccount() {
+        return super.defaultAccount;
+    }
+
+    /**
+     * Setter for the defaultBlock property
+     *
+     * @property defaultBlock
+     *
+     * @param {String|Number}value
+     */
+    set defaultBlock(value) {
+        if (this.resolverContract) {
+            this.resolverContract.defaultBlock = value;
+        }
+
+        super.defaultBlock = value;
+    }
+
+    /**
+     * Gets the defaultBlock property
+     *
+     * @property defaultBlock
+     *
+     * @returns {String|Number} value
+     */
+    get defaultBlock() {
+        return super.defaultBlock;
+    }
+
+    /**
      * Returns the address of the owner of an Ens name.
      *
      * @method owner
@@ -89,7 +314,7 @@ export default class Registry extends AbstractContract {
      */
     setProvider(provider, net) {
         if (this.resolverContract) {
-            return this.resolverContract.setProvider(provider, net) && super.setProvider(provider, net);
+            this.resolverContract.setProvider(provider, net);
         }
 
         return super.setProvider(provider, net);