+++ packages/web3-eth/types/tests/eth.tests.ts
@@ -44,6 +44,9 @@ eth.abi;
 // $ExpectType Network
 eth.net;
 
+// $ExpectType TransactionSigner
+eth.transactionSigner;
+
 eth.clearSubscriptions();
 
 // $ExpectType Subscription<Log>