+++ CHANGELOG.md
@@ -91,7 +91,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.0-alpha.2]
 
+### Added
+
+- ``automine`` and ``Instant Seal`` support added (#2940)
+
+### Changed
+
+- default value of ``transactionConfirmationBlocks`` changed to ``0`` (#3008)
+- ``TransactionObserver`` splitted up to ``HttpTransactionObserver`` and ``SocketTransactionObserver`` (#3008)
+
 ### Fixed
 
 - Error handling of the HttpProvider fixed (#2887)
+- Property handling for the Ens module fixed (#3008)